New org perms (#102)
* add OrgRole to db & create pages

* update seed

* update seed, move owner attribute to org user rather than org

* refactor to support new org permissions system

* format

* start of ui for org roles

* update values

* format

* ui fixes

* color button right side

* permissions UI

* refactor to use descriptions

* saving of permissions

* toasts are neat

* format

* delete roles

* one of the migrations of all time

* That was NOT one of the migrations of all time

* update seed

* allow assigning roles

* searchbox changes

* update creating club permissions object

* allow org admins to manage roles in clubs

* more club permission fixes

* fix some small issues w/ org perm to club perm logic

* allow org admins to change club ownership

* format

* refactor: unify common perm files

* refactor: shorten duplicate info

* fix: nitpicks

* style: fmt

* fix: check

* fix bug that allowed anyone to delete orgs

* added view settings perm for club users

* allow removing roles and fix searchbox issue

* fix small perms issue with club roles

* format

---------

+++ src/routes/(app)/org/[id=integer]/settings/danger/+page.server.ts
@@ -1,5 +1,6 @@
 import { redirect } from '@sveltejs/kit';
 
+import { createOrgPermissionsFromUser } from '$lib/permissions/orgPermissions.js';
 import { prisma } from '$lib/server/prismaConnection.js';
 import { verifySession } from '$lib/server/verifySession.js';
 
@@ -28,7 +29,27 @@ export const load = async ({ parent }) => {
 
 export const actions = {
 	refreshJoinCode: async ({ cookies, params }) => {
-		await verifySession(cookies.get('session'));
+		const user = await verifySession(cookies.get('session'), {
+			orgUsers: {
+				include: {
+					role: true
+				}
+			}
+		});
+		const org = await prisma.organization.findFirst({
+			where: {
+				id: parseInt(params.id)
+			}
+		});
+
+		const perms = createOrgPermissionsFromUser(user, org);
+
+		if (!perms.admin) {
+			return {
+				success: false,
+				message: 'Only admins can refresh join code.'
+			};
+		}
 
 		// TODO: better join code generation
 		const orgAmount = await prisma.organization.count();
@@ -50,7 +71,22 @@ export const actions = {
 		};
 	},
 	deleteOrg: async ({ cookies, params }) => {
-		await verifySession(cookies.get('session'));
+		const user = await verifySession(cookies.get('session'));
+		const orgUser = await prisma.orgUser.findFirst({
+			where: {
+				AND: {
+					organizationId: parseInt(params.id),
+					userId: user.id
+				}
+			}
+		});
+
+		if (!orgUser?.owner) {
+			return {
+				success: false,
+				message: 'Only an owner can delete an organization.'
+			};
+		}
 
 		await prisma.organization.delete({
 			where: {
